requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings

langchain
langchain-core
langchain-openai
langchain-chroma
chromadb

python-dotenv